Modificações na no controller e service de Anime

+++ src/main/java/duarte/br/desafioapirestdio/service/AnimesService.java
@@ -1,14 +1,21 @@
 package duarte.br.desafioapirestdio.service;
 
+import duarte.br.desafioapirestdio.dto.AnimesDto;
 import duarte.br.desafioapirestdio.model.Animes;
 import duarte.br.desafioapirestdio.repository.AnimesRepository;
+import duarte.br.desafioapirestdio.service.exceptions.DataBaseException;
+import duarte.br.desafioapirestdio.service.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class AnimesService {
@@ -16,36 +23,64 @@ public class AnimesService {
     @Autowired
     private AnimesRepository animesRepository;
 
-    public Page<Animes> listAll(Pageable pageable) {
-        return animesRepository.findAll(pageable);
+    @Transactional(readOnly = true)
+    public Page<AnimesDto> listAll(Pageable pageable) {
+        Page<Animes> animesPage = animesRepository.findAll(pageable);
+        return animesPage.map(this::convertToDto);
     }
 
-    public List<Animes> findByName(String title) {
-        return animesRepository.findByTitle(title);
+    @Transactional(readOnly = true)
+    public List<AnimesDto> findByName(String title) {
+        List<Animes> animes = animesRepository.findByTitle(title);
+
+        return animes.stream().map(this::convertToDto).collect(Collectors.toList());
     }
 
-    public Animes save(Animes animes) {
-        return animesRepository.save(animes);
+    @Transactional
+    public AnimesDto save(AnimesDto animesDto) {
+        Animes entity = new Animes();
+        copyDtoToEntity(animesDto, entity);
+        entity = animesRepository.save(entity);
+        return convertToDto(entity);
+
     }
 
-    public Optional<Animes> update(Long id, Animes anime) {
-        Optional<Animes> animeUpdate = animesRepository.findById(id);
+    @Transactional
+    public AnimesDto update(Long id, AnimesDto animesDto) {
+        try {
+            Animes entity = animesRepository.getReferenceById(id);
+            copyDtoToEntity(animesDto, entity);
+            entity = animesRepository.save(entity);
+            return new AnimesDto(entity);
 
-        if (animeUpdate.isPresent()) {
-            Animes existingAnime = animeUpdate.get();
-            existingAnime.setTitle(anime.getTitle());
-            existingAnime.setStudio(anime.getStudio());
-            existingAnime.setDirector(anime.getDirector());
-            existingAnime.setDescription(anime.getDescription());
-
-            return Optional.of(animesRepository.save(existingAnime));
-        } else {
-            return Optional.empty();
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
         }
     }
 
+    @Transactional(propagation = Propagation.SUPPORTS)
 
     public void delete(Long id) {
-        animesRepository.deleteById(id);
+        if (!animesRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            animesRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DataBaseException("Falha na integridade referencial");
+        }
+    }
+
+    private AnimesDto convertToDto(Animes animes) {
+        return new AnimesDto(animes.getId(), animes.getTitle(), animes.getDescription(), animes.getImageUrl(),
+                animes.getDirectors(), animes.getStudios());
+    }
+
+    private void copyDtoToEntity(AnimesDto dto, Animes entity) {
+        entity.setTitle(dto.getTitle());
+        entity.setDescription(dto.getDescription());
+        entity.setImageUrl(dto.getImageUrl());
+        entity.setStudios(dto.getStudios());
+        entity.setDirectors(dto.getDirectors());
     }
 }
